+++ telegram-bot/src/commonMain/kotlin/eu/vendeli/tgbot/types/internal/ActivitiesData.kt
@@ -1,5 +1,6 @@
 package eu.vendeli.tgbot.types.internal
 
+import eu.vendeli.tgbot.annotations.internal.InternalApi
 import eu.vendeli.tgbot.core.TgUpdateHandler
 import eu.vendeli.tgbot.utils.CommandHandlers
 import eu.vendeli.tgbot.utils.CommonHandlers
@@ -13,6 +14,7 @@ import eu.vendeli.tgbot.utils.logString
 internal class ActivitiesData(
     pkg: String? = null,
 ) {
+    @OptIn(InternalApi::class)
     private val activities = when {
         _OperatingActivities.size == 1 -> _OperatingActivities.entries.firstOrNull()?.value
         _OperatingActivities.size > 1 && pkg != null -> _OperatingActivities[pkg]