+++ webapps/src/jsMain/kotlin/eu/vendeli/webapps/utils/WebAppExtUtils.kt
@@ -10,63 +10,65 @@ val WebApp.ColorScheme: ColorScheme
     get() = eu.vendeli.webapps.ui.ColorScheme.valueOf(colorScheme)
 
 sealed interface EventHandler {
-    value class ViewportChanged(val block: ViewportChangedEventHandler) : EventHandler
-    value class InvoiceClosed(val block: InvoiceClosedEventHandler) : EventHandler
-    value class PopupClosed(val block: PopupClosedEventHandler) : EventHandler
-    value class QRTextReceived(val block: QRTextReceivedEventHandler) : EventHandler
-    value class ClipboardTextReceived(val block: ClipboardTextReceivedEventHandler) : EventHandler
-    value class WriteAccessRequested(val block: WriteAccessRequestedHandler) : EventHandler
-    value class ContactRequested(val block: ContactRequestedHandler) : EventHandler
-    value class NoParams(val block: NoParamsEventHandler) : EventHandler
+    val block: Function<*>
+
+    value class ViewportChanged(override val block: ViewportChangedEventHandler) : EventHandler
+    value class InvoiceClosed(override val block: InvoiceClosedEventHandler) : EventHandler
+    value class PopupClosed(override val block: PopupClosedEventHandler) : EventHandler
+    value class QRTextReceived(override val block: QRTextReceivedEventHandler) : EventHandler
+    value class ClipboardTextReceived(override val block: ClipboardTextReceivedEventHandler) : EventHandler
+    value class WriteAccessRequested(override val block: WriteAccessRequestedHandler) : EventHandler
+    value class ContactRequested(override val block: ContactRequestedHandler) : EventHandler
+    value class NoParams(override val block: NoParamsEventHandler) : EventHandler
 }
 
 fun WebApp.onEvent(type: EventType.ViewportChanged, eventHandler: EventHandler.ViewportChanged) {
-    onEvent(type.name, eventHandler)
+    onEvent(type.name, eventHandler.block)
 }
 
 fun WebApp.onEvent(type: EventType.InvoiceClosed, eventHandler: EventHandler.InvoiceClosed) {
-    onEvent(type.name, eventHandler)
+    onEvent(type.name, eventHandler.block)
 }
 
 fun WebApp.onEvent(type: EventType.PopupClosed, eventHandler: EventHandler.PopupClosed) {
-    onEvent(type.name, eventHandler)
+    onEvent(type.name, eventHandler.block)
 }
 
 fun WebApp.onEvent(type: EventType.QRTextReceived, eventHandler: EventHandler.QRTextReceived) {
-    onEvent(type.name, eventHandler)
+    onEvent(type.name, eventHandler.block)
 }
 
 fun WebApp.onEvent(type: EventType.ClipboardTextReceived, eventHandler: EventHandler.ClipboardTextReceived) {
-    onEvent(type.name, eventHandler)
+    onEvent(type.name, eventHandler.block)
 }
 
 fun WebApp.onEvent(type: EventType.WriteAccessRequested, eventHandler: EventHandler.WriteAccessRequested) {
-    onEvent(type.name, eventHandler)
+    onEvent(type.name, eventHandler.block)
 }
 
 fun WebApp.onEvent(type: EventType.ContactRequested, eventHandler: EventHandler.ContactRequested) {
-    onEvent(type.name, eventHandler)
+    onEvent(type.name, eventHandler.block)
 }
 
 // no param events
 fun WebApp.onEvent(type: EventType.ThemeChanged, eventHandler: EventHandler.NoParams) {
-    onEvent(type.name, eventHandler)
+    onEvent(type.name, eventHandler.block)
 }
 
 fun WebApp.onEvent(type: EventType.MainButtonClicked, eventHandler: EventHandler.NoParams) {
-    onEvent(type.name, eventHandler)
+    onEvent(type.name, eventHandler.block)
 }
 
 fun WebApp.onEvent(type: EventType.BackButtonClicked, eventHandler: EventHandler.NoParams) {
-    onEvent(type.name, eventHandler)
+    onEvent(type.name, eventHandler.block)
 }
 
 fun WebApp.onEvent(type: EventType.SettingsButtonClicked, eventHandler: EventHandler.NoParams) {
-    onEvent(type.name, eventHandler)
+    onEvent(type.name, eventHandler.block)
 }
 
 fun WebApp.onEvent(type: EventType.ScanQrPopupClosed, eventHandler: EventHandler.NoParams) {
-    onEvent(type.name, eventHandler)
+    onEvent(type.name, eventHandler.block)
 }
 
 fun WebApp.showPopup(